Unified test files of quickfix tests
