+++ server/src/test/kotlin/org/javacs/kt/QuickFixesTest.kt
@@ -3,8 +3,9 @@ package org.javacs.kt
 import org.eclipse.lsp4j.CodeActionKind
 import org.eclipse.lsp4j.Diagnostic
 import org.eclipse.lsp4j.jsonrpc.messages.Either
-import org.hamcrest.Matchers
-import org.junit.Assert
+import org.hamcrest.Matchers.equalTo
+import org.hamcrest.Matchers.hasSize
+import org.junit.Assert.assertThat
 import org.junit.Test
 
 class ImplementAbstractFunctionsQuickFixTest : SingleFileTestFixture("quickfixes", "SomeSubclass.kt") {
@@ -16,27 +17,27 @@ class ImplementAbstractFunctionsQuickFixTest : SingleFileTestFixture("quickfixes
 
         val codeActions = languageServer.textDocumentService.codeAction(codeActionParams).get()
 
-        Assert.assertThat(codeActions.size, Matchers.equalTo(1))
-        Assert.assertThat(codeActions[0].right.kind, Matchers.equalTo(CodeActionKind.QuickFix))
-        Assert.assertThat(codeActions[0].right.diagnostics.size, Matchers.equalTo(1))
-        Assert.assertThat(codeActions[0].right.diagnostics[0], Matchers.equalTo(diagnostic))
-        Assert.assertThat(codeActions[0].right.edit.changes.size, Matchers.equalTo(1))
-        Assert.assertThat(codeActions[0].right.edit.changes[codeActionParams.textDocument.uri]?.size, Matchers.equalTo(2))
-        Assert.assertThat(
+        assertThat(codeActions.size, equalTo(1))
+        assertThat(codeActions[0].right.kind, equalTo(CodeActionKind.QuickFix))
+        assertThat(codeActions[0].right.diagnostics.size, equalTo(1))
+        assertThat(codeActions[0].right.diagnostics[0], equalTo(diagnostic))
+        assertThat(codeActions[0].right.edit.changes.size, equalTo(1))
+        assertThat(codeActions[0].right.edit.changes[codeActionParams.textDocument.uri]?.size, equalTo(2))
+        assertThat(
             codeActions[0].right.edit.changes[codeActionParams.textDocument.uri]?.get(0)?.range,
-            Matchers.equalTo(range(3, 55, 3, 55))
+            equalTo(range(3, 55, 3, 55))
         )
-        Assert.assertThat(
+        assertThat(
             codeActions[0].right.edit.changes[codeActionParams.textDocument.uri]?.get(0)?.newText,
-            Matchers.equalTo(System.lineSeparator() + System.lineSeparator() + "    override fun someSuperMethod(someParameter: String): Int { }")
+            equalTo(System.lineSeparator() + System.lineSeparator() + "    override fun someSuperMethod(someParameter: String): Int { }")
         )
-        Assert.assertThat(
+        assertThat(
             codeActions[0].right.edit.changes[codeActionParams.textDocument.uri]?.get(1)?.range,
-            Matchers.equalTo(range(3, 55, 3, 55))
+            equalTo(range(3, 55, 3, 55))
         )
-        Assert.assertThat(
+        assertThat(
             codeActions[0].right.edit.changes[codeActionParams.textDocument.uri]?.get(1)?.newText,
-            Matchers.equalTo(System.lineSeparator() + System.lineSeparator() + "    override fun someInterfaceMethod() { }")
+            equalTo(System.lineSeparator() + System.lineSeparator() + "    override fun someInterfaceMethod() { }")
         )
     }
 
@@ -48,19 +49,19 @@ class ImplementAbstractFunctionsQuickFixTest : SingleFileTestFixture("quickfixes
 
         val codeActions = languageServer.textDocumentService.codeAction(codeActionParams).get()
 
-        Assert.assertThat(codeActions.size, Matchers.equalTo(1))
-        Assert.assertThat(codeActions[0].right.kind, Matchers.equalTo(CodeActionKind.QuickFix))
-        Assert.assertThat(codeActions[0].right.diagnostics.size, Matchers.equalTo(1))
-        Assert.assertThat(codeActions[0].right.diagnostics[0], Matchers.equalTo(diagnostic))
-        Assert.assertThat(codeActions[0].right.edit.changes.size, Matchers.equalTo(1))
-        Assert.assertThat(codeActions[0].right.edit.changes[codeActionParams.textDocument.uri]?.size, Matchers.equalTo(1))
-        Assert.assertThat(
+        assertThat(codeActions.size, equalTo(1))
+        assertThat(codeActions[0].right.kind, equalTo(CodeActionKind.QuickFix))
+        assertThat(codeActions[0].right.diagnostics.size, equalTo(1))
+        assertThat(codeActions[0].right.diagnostics[0], equalTo(diagnostic))
+        assertThat(codeActions[0].right.edit.changes.size, equalTo(1))
+        assertThat(codeActions[0].right.edit.changes[codeActionParams.textDocument.uri]?.size, equalTo(1))
+        assertThat(
             codeActions[0].right.edit.changes[codeActionParams.textDocument.uri]?.get(0)?.range,
-            Matchers.equalTo(range(7, 74, 7, 74))
+            equalTo(range(7, 74, 7, 74))
         )
-        Assert.assertThat(
+        assertThat(
             codeActions[0].right.edit.changes[codeActionParams.textDocument.uri]?.get(0)?.newText,
-            Matchers.equalTo(System.lineSeparator() + System.lineSeparator() + "    override fun someInterfaceMethod() { }")
+            equalTo(System.lineSeparator() + System.lineSeparator() + "    override fun someInterfaceMethod() { }")
         )
     }
 
@@ -72,19 +73,170 @@ class ImplementAbstractFunctionsQuickFixTest : SingleFileTestFixture("quickfixes
 
         val codeActions = languageServer.textDocumentService.codeAction(codeActionParams).get()
 
-        Assert.assertThat(codeActions.size, Matchers.equalTo(1))
-        Assert.assertThat(codeActions[0].right.kind, Matchers.equalTo(CodeActionKind.QuickFix))
-        Assert.assertThat(codeActions[0].right.diagnostics.size, Matchers.equalTo(1))
-        Assert.assertThat(codeActions[0].right.diagnostics[0], Matchers.equalTo(diagnostic))
-        Assert.assertThat(codeActions[0].right.edit.changes.size, Matchers.equalTo(1))
-        Assert.assertThat(codeActions[0].right.edit.changes[codeActionParams.textDocument.uri]?.size, Matchers.equalTo(1))
-        Assert.assertThat(
+        assertThat(codeActions.size, equalTo(1))
+        assertThat(codeActions[0].right.kind, equalTo(CodeActionKind.QuickFix))
+        assertThat(codeActions[0].right.diagnostics.size, equalTo(1))
+        assertThat(codeActions[0].right.diagnostics[0], equalTo(diagnostic))
+        assertThat(codeActions[0].right.edit.changes.size, equalTo(1))
+        assertThat(codeActions[0].right.edit.changes[codeActionParams.textDocument.uri]?.size, equalTo(1))
+        assertThat(
             codeActions[0].right.edit.changes[codeActionParams.textDocument.uri]?.get(0)?.range,
-            Matchers.equalTo(range(11, 43, 11, 43))
+            equalTo(range(11, 43, 11, 43))
         )
-        Assert.assertThat(
+        assertThat(
             codeActions[0].right.edit.changes[codeActionParams.textDocument.uri]?.get(0)?.newText,
-            Matchers.equalTo(System.lineSeparator() + System.lineSeparator() + "    override fun someSuperMethod(someParameter: String): Int { }")
+            equalTo(System.lineSeparator() + System.lineSeparator() + "    override fun someSuperMethod(someParameter: String): Int { }")
         )
+    }
+}
+
+class ImplementAbstractFunctionsQuickFixSameFileTest : SingleFileTestFixture("quickfixes", "samefile.kt") {
+    @Test
+    fun `should find no code actions`() {
+        val only = listOf(CodeActionKind.QuickFix)
+        val codeActionParams = codeActionParams(file, 3, 1, 3, 22, diagnostics, only)
+    
+        val codeActionResult = languageServer.textDocumentService.codeAction(codeActionParams).get()
+
+        assertThat(codeActionResult, hasSize(0))
+    }
+
+    @Test
+    fun `should find one abstract method to implement`() {
+        val only = listOf(CodeActionKind.QuickFix)
+        val codeActionParams = codeActionParams(file, 7, 1, 7, 14, diagnostics, only)
+    
+        val codeActionResult = languageServer.textDocumentService.codeAction(codeActionParams).get()
+
+        assertThat(codeActionResult, hasSize(1))
+        val codeAction = codeActionResult[0].right
+        assertThat(codeAction.kind, equalTo(CodeActionKind.QuickFix))
+        assertThat(codeAction.title, equalTo("Implement abstract functions"))
+        assertThat(codeAction.diagnostics, equalTo(listOf(diagnostics[0])))
+
+        val textEdit = codeAction.edit.changes
+        val key = workspaceRoot.resolve(file).toUri().toString()
+        assertThat(textEdit.containsKey(key), equalTo(true))
+        assertThat(textEdit[key], hasSize(1))
+
+        val functionToImplementEdit = textEdit[key]?.get(0)
+        assertThat(functionToImplementEdit?.range, equalTo(range(7, 30, 7, 30)))
+        assertThat(functionToImplementEdit?.newText, equalTo(System.lineSeparator() + System.lineSeparator() + "    override fun test(input: String, otherInput: Int) { }"))
+    }
+
+    @Test
+    fun `should find several abstract methods to implement`() {
+        val only = listOf(CodeActionKind.QuickFix)
+        val codeActionParams = codeActionParams(file, 15, 1, 15, 21, diagnostics, only)
+    
+        val codeActionResult = languageServer.textDocumentService.codeAction(codeActionParams).get()
+
+        assertThat(codeActionResult, hasSize(1))
+        val codeAction = codeActionResult[0].right
+        assertThat(codeAction.kind, equalTo(CodeActionKind.QuickFix))
+        assertThat(codeAction.title, equalTo("Implement abstract functions"))
+
+        val textEdit = codeAction.edit.changes
+        val key = workspaceRoot.resolve(file).toUri().toString()
+        assertThat(textEdit.containsKey(key), equalTo(true))
+        assertThat(textEdit[key], hasSize(2))
+
+        val firstFunctionToImplementEdit = textEdit[key]?.get(0)
+        assertThat(firstFunctionToImplementEdit?.range, equalTo(range(15, 49, 15, 49)))
+        assertThat(firstFunctionToImplementEdit?.newText, equalTo(System.lineSeparator() + System.lineSeparator() + "    override fun print() { }"))
+        
+        val secondFunctionToImplementEdit = textEdit[key]?.get(1)
+        assertThat(secondFunctionToImplementEdit?.range, equalTo(range(15, 49, 15, 49)))
+        assertThat(secondFunctionToImplementEdit?.newText, equalTo(System.lineSeparator() + System.lineSeparator() + "    override fun test(input: String, otherInput: Int) { }"))
+    }
+
+    @Test
+    fun `should find only one abstract method when the other one is already implemented`() {
+        val only = listOf(CodeActionKind.QuickFix)
+        val codeActionParams = codeActionParams(file, 17, 1, 17, 26, diagnostics, only)
+    
+        val codeActionResult = languageServer.textDocumentService.codeAction(codeActionParams).get()
+
+        assertThat(codeActionResult, hasSize(1))
+        val codeAction = codeActionResult[0].right
+        assertThat(codeAction.kind, equalTo(CodeActionKind.QuickFix))
+        assertThat(codeAction.title, equalTo("Implement abstract functions"))
+
+        val textEdit = codeAction.edit.changes
+        val key = workspaceRoot.resolve(file).toUri().toString()
+        assertThat(textEdit.containsKey(key), equalTo(true))
+        assertThat(textEdit[key], hasSize(1))
+        
+        val functionToImplementEdit = textEdit[key]?.get(0)
+        assertThat(functionToImplementEdit?.range, equalTo(range(18, 57, 18, 57)))
+        assertThat(functionToImplementEdit?.newText, equalTo(System.lineSeparator() + System.lineSeparator() + "    override fun print() { }"))
+    }
+
+    @Test
+    fun `should respect nullability of parameter and return value in abstract method`() {
+        val only = listOf(CodeActionKind.QuickFix)
+        val codeActionParams = codeActionParams(file, 25, 1, 25, 16, diagnostics, only)
+    
+        val codeActionResult = languageServer.textDocumentService.codeAction(codeActionParams).get()
+
+        assertThat(codeActionResult, hasSize(1))
+        val codeAction = codeActionResult[0].right
+        assertThat(codeAction.kind, equalTo(CodeActionKind.QuickFix))
+        assertThat(codeAction.title, equalTo("Implement abstract functions"))
+
+        val textEdit = codeAction.edit.changes
+        val key = workspaceRoot.resolve(file).toUri().toString()
+        assertThat(textEdit.containsKey(key), equalTo(true))
+        assertThat(textEdit[key], hasSize(1))
+
+        val functionToImplementEdit = textEdit[key]?.get(0)
+        assertThat(functionToImplementEdit?.range, equalTo(range(25, 48, 25, 48)))
+        assertThat(functionToImplementEdit?.newText, equalTo(System.lineSeparator() + System.lineSeparator() + "    override fun myMethod(myStr: String?): String? { }"))
+    }
+}
+
+class ImplementAbstractFunctionsQuickFixExternalLibraryTest : SingleFileTestFixture("quickfixes", "standardlib.kt") {
+    @Test
+    fun `should find one abstract method from Runnable to implement`() {
+        val only = listOf(CodeActionKind.QuickFix)
+        val codeActionParams = codeActionParams(file, 5, 1, 5, 15, diagnostics, only)
+    
+        val codeActionResult = languageServer.textDocumentService.codeAction(codeActionParams).get()
+
+        assertThat(codeActionResult, hasSize(1))
+        val codeAction = codeActionResult[0].right
+        assertThat(codeAction.kind, equalTo(CodeActionKind.QuickFix))
+        assertThat(codeAction.title, equalTo("Implement abstract functions"))
+
+        val textEdit = codeAction.edit.changes
+        val key = workspaceRoot.resolve(file).toUri().toString()
+        assertThat(textEdit.containsKey(key), equalTo(true))
+        assertThat(textEdit[key], hasSize(1))
+
+        val functionToImplementEdit = textEdit[key]?.get(0)
+        assertThat(functionToImplementEdit?.range, equalTo(range(5, 28, 5, 28)))
+        assertThat(functionToImplementEdit?.newText, equalTo(System.lineSeparator() + System.lineSeparator() + "    override fun run() { }"))
+    }
+
+    @Test
+    fun `should find one abstract method from Comparable to implement`() {
+        val only = listOf(CodeActionKind.QuickFix)
+        val codeActionParams = codeActionParams(file, 7, 1, 7, 19, diagnostics, only)
+    
+        val codeActionResult = languageServer.textDocumentService.codeAction(codeActionParams).get()
+
+        assertThat(codeActionResult, hasSize(1))
+        val codeAction = codeActionResult[0].right
+        assertThat(codeAction.kind, equalTo(CodeActionKind.QuickFix))
+        assertThat(codeAction.title, equalTo("Implement abstract functions"))
+
+        val textEdit = codeAction.edit.changes
+        val key = workspaceRoot.resolve(file).toUri().toString()
+        assertThat(textEdit.containsKey(key), equalTo(true))
+        assertThat(textEdit[key], hasSize(1))
+
+        val functionToImplementEdit = textEdit[key]?.get(0)
+        assertThat(functionToImplementEdit?.range, equalTo(range(7, 42, 7, 42)))
+        assertThat(functionToImplementEdit?.newText, equalTo(System.lineSeparator() + System.lineSeparator() + "    override fun compare(p0: String, p1: String): Int { }"))
     }
 }
